[llvm-bcanalyzer] Refactor and move to libLLVMBitReader
This allows us to use the analyzer from unit tests.

* Refactor the interface to use proper error handling for most functions
  after JF's work.
* Move everything into a BitstreamAnalyzer class.
* Move that to Bitcode/BitcodeAnalyzer.h.

Differential Revision: https://reviews.llvm.org/D64116

llvm-svn: 365286
(cherry picked from commit 4cdb68ebbd52ea536dcd16ad10402d4827a7bbe6)

 Conflicts:
	llvm/tools/llvm-bcanalyzer/llvm-bcanalyzer.cpp
